added example for custom modifier

+++ contracts/7_ModifierExample.sol
@@ -2,32 +2,37 @@
 
 pragma solidity ^0.8.7;
 
-contract ModifierExample{
+contract ModifierExample {
     uint256 public userCount = 0;
     address owner;
 
-    modifier  onlyOwner () {
-        require (msg.sender==owner);
+    modifier onlyOwner() {
+        require(msg.sender == owner);
         _;
     }
 
-    mapping (uint => User ) public Users;
-    struct User{
-        uint _id;
+    mapping(uint256 => User) public Users;
+    struct User {
+        uint256 _id;
         string _name;
-        uint256 _balance;        
+        uint256 _balance;
     }
 
     constructor() {
         owner = msg.sender;
-    }    
+    }
 
     function incrementUserCount() internal {
         userCount += 1;
     }
 
-    function addUser(string memory _name, uint256 _amt) public onlyOwner {
-        incrementUserCount();
-        Users[userCount] = User(userCount, _name, _amt);
+    function addUser(
+        string memory _name, 
+        uint256 _amt
+        ) 
+        public onlyOwner 
+        {
+            incrementUserCount();
+            Users[userCount] = User(userCount, _name, _amt);
     }
 }
